# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUSFORGE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HubStatus {
  status: string;
  version: string;
  timestamp: string;
  engines: string[];
}

export interface EngineMap {
  php: boolean;
  node: boolean;
  python: boolean;
  java: boolean;
  cpp: boolean;
  powershell: boolean;
  bash: boolean;
}

export interface ExecInput {
  cmd: string;
  engine: string;
  mode?: string;
  attacker?: string;
}

export interface ExecResult {
  output: string;
  engine: string;
  mode: string;
  elapsed: number;
}

export interface InjectionLog {
  id: number;
  timestamp: string;
  command: string;
  engine: string;
  mode: string;
  responseTime: number;
}

export interface ClearResult {
  ok: boolean;
}

export type GetSuggestionsParams = {
mode?: string;
cmd?: string;
};