# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUSFORGE API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hub/exec:
    post:
      operationId: execPayload
      tags: [hub]
      summary: Execute a payload through the selected engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecInput"
      responses:
        "200":
          description: Execution output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResult"
        "400":
          description: Bad request
...
components:
  schemas:
...
    ExecInput:
      type: object
      properties:
        cmd:
          type: string
        engine:
          type: string
        mode:
          type: string
        attacker:
          type: string
      required:
        - cmd
        - engine
    ExecResult:
      type: object
      properties:
        output:
          type: string
        engine:
          type: string
        mode:
          type: string
        elapsed:
          type: integer
      required:
        - output
        - engine
        - mode
        - elapsed
...